modify 首次安装配置文件备份、sql EXception

+++ rasp-install/java/src/main/java/com/baidu/rasp/install/BaseStandardInstaller.java
@@ -50,6 +50,7 @@ public abstract class BaseStandardInstaller implements Installer {
 
     @Override
     public void install() throws RaspError, IOException {
+        boolean firstInstall = false;
         String jarPath = getClass().getProtectionDomain().getCodeSource().getLocation().getPath();
         File srcDir = new File(new File(jarPath).getParent() + File.separator + "rasp");
         if (!(srcDir.exists() && srcDir.isDirectory())) {
@@ -57,6 +58,10 @@ public abstract class BaseStandardInstaller implements Installer {
         }
         File installDir = new File(getInstallPath(serverRoot));
 
+        File configFile = new File(installDir.getCanonicalPath() + File.separator + "conf" + File.separator + "rasp.yaml");
+        if (!configFile.exists()) {
+            firstInstall = true;
+        }
         if (!srcDir.getCanonicalPath().equals(installDir.getCanonicalPath())) {
             // 拷贝rasp文件夹
             System.out.println("Duplicating \"rasp\" directory\n- " + installDir.getCanonicalPath());
@@ -67,7 +72,7 @@ public abstract class BaseStandardInstaller implements Installer {
         modifyFolerPermission(installDir.getCanonicalPath());
 
         // 生成配置文件
-        if (!generateConfig(installDir.getPath())) {
+        if (!generateConfig(installDir.getPath(), firstInstall)) {
             System.exit(1);
         }
 
@@ -86,7 +91,7 @@ public abstract class BaseStandardInstaller implements Installer {
     }
 
 
-    private boolean generateConfig(String dir) {
+    private boolean generateConfig(String dir, boolean firstInstall) {
         try {
             String sep = File.separator;
             File target = new File(dir + sep + "conf" + sep + "rasp.yaml");
@@ -96,10 +101,10 @@ public abstract class BaseStandardInstaller implements Installer {
                 System.out.println("- Already exists and reserved rasp.yaml, continuing ..");
                 return true;
             }
-            if (target.exists()) {
+            if (target.exists() && !firstInstall) {
                 File reserve = new File(dir + sep + "conf" + sep + "rasp.yaml.bak");
                 if (!reserve.exists()) {
-                  reserve.createNewFile();
+                    reserve.createNewFile();
                 }
                 FileOutputStream outputStream = new FileOutputStream(reserve);
                 FileInputStream inputStream = new FileInputStream(target);
@@ -178,13 +183,13 @@ public abstract class BaseStandardInstaller implements Installer {
             if (url != null && appId != null && appSecret != null) {
                 String path = getInstallPath(serverRoot) + File.separator + "conf" + File.separator + "rasp.yaml";
                 File yamlFile = new File(path);
-                if (yamlFile.exists()){
-                    Map<String,Object> map = new HashMap<String, Object>();
+                if (yamlFile.exists()) {
+                    Map<String, Object> map = new HashMap<String, Object>();
                     map.put("cloud.enable", true);
                     map.put("cloud.backend_url", url);
                     map.put("cloud.app_id", appId);
                     map.put("cloud.app_secret", appSecret);
-                    FileWriter writer = new FileWriter(yamlFile,true);
+                    FileWriter writer = new FileWriter(yamlFile, true);
                     writer.write(LINE_SEP);
                     writer.write("#云控配置");
                     writer.write(LINE_SEP);
@@ -192,7 +197,7 @@ public abstract class BaseStandardInstaller implements Installer {
                     options.setDefaultFlowStyle(DumperOptions.FlowStyle.BLOCK);
                     options.setPrettyFlow(true);
                     Yaml yaml = new Yaml(options);
-                    yaml.dump(map,writer);
+                    yaml.dump(map, writer);
                 }
             }
         } catch (Exception e) {
@@ -201,7 +206,7 @@ public abstract class BaseStandardInstaller implements Installer {
     }
 
     //判断tomcat的版本是否大于8
-    protected boolean checkTomcatVersion(){
+    protected boolean checkTomcatVersion() {
         String javaVersion = System.getProperty("java.version");
         return javaVersion != null && (javaVersion.startsWith("1.9") || javaVersion.startsWith("10.")
                 || javaVersion.startsWith("11."));
